requirements.txt
langchain~=0.3.27
langchain-anthropic~=0.3.22
langchain-openai~=0.3.16
python-dotenv~=0.21.0
langchain-ollama~=0.3.10
langchain-tavily~=0.2.13
tavily-python~=0.7.14

langchain-core~=1.1.0
requests~=2.32.5